Adds error helpers.
Error helpers wrap a region with an implicit single execution
do/while, and provides a number of test macros to ease checking
for error conditions, setting some error flag, and breaking out
of the region.  This allows cleanup code to be gathered outside
the end of the error region rather than having lots of fragile
checks within the main logic.

This allowed removing a number of state types, which had been
used to similar effect previously.