Adds next phase of cue traversal process testing.
We know validate that the cue files and structure of the target
directory is as expected.  The cues will be transformed, though
we don't verify that (verified in separate test).  Directories
without cues are pruned.

We delete the target dir at the end of the test.

We still need to copy over files reference by the cue, both those
that should be directly copied, and those that will be
converted during the process.

We should also report whether the directory creation and cue
writing fails.