Fixes non-golden path for cue convert/write errors.
Previously, a method in the error chain was eating all errors,
returning success, so the error cases weren't being reported
properly.  This is now fix, and the error output has been
verified.

This can be verified by creating the directory the cue files
will be transformed into in the test_cue_traverse test, and
removing all permissions from it.  The writes into the
directory will fail, generating the error reports.