Starts work on duplicate copying (with overwrite).
The converter should be able to skip previously converted output,
just converting newly discovered cues, but should also be able to
be forced to overwrite everything.

This adds the existence check, but not yet the overwrite or any
tests.

Test work has started but a directory copy routine is needed for
setting up the environment.  This change starts extracting some
of the parallel traversal logic into a separate helper that should
be usable for sirectory copy.  The existing parallel traverse test
was updated to use the new visitor, but it has not been used
within cue convert itself.